fix duts location in result object
